#101 이미지 캐시 구현
